% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_fixer.R
\name{name_fixer}
\alias{name_fixer}
\title{Convert name inputs into working URL paths}
\usage{
name_fixer(x)
}
\arguments{
\item{x}{Name to convert}
}
\description{
Convert name inputs into working URL paths
}
