% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rider_profile.R
\name{parse_rider_profile}
\alias{parse_rider_profile}
\title{Parse rider profile information from HTML code}
\usage{
parse_rider_profile(rider_html)
}
\arguments{
\item{rider_html}{HTML code of rider's profile page}
}
\value{
Rider profile information (see \code{rider_profiles_men} documentation)
}
\description{
\code{parse_rider_profile} parses HTML code of rider's profile page
for personal information.
}
