% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rider_results.R
\name{parse_rider_results}
\alias{parse_rider_results}
\title{Parse rider results from HTML code}
\usage{
parse_rider_results(rider_id, rider_html, seasons = NULL)
}
\arguments{
\item{rider_id}{Rider's profile ID}

\item{rider_html}{HTML code of rider's profile page}

\item{seasons}{(optional) Vector of covered seasons (as integers), all if \code{NULL} (default)}
}
\value{
Rider results (see \code{rider_records_men} documentation)
}
\description{
\code{parse_rider_results} parses HTML code of rider's profile page
for race results.
}
