% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_team.R
\name{parse_team}
\alias{parse_team}
\title{Parse team members from HTML}
\usage{
parse_team(team_url, team)
}
\arguments{
\item{team_url}{team name}
}
\value{
A tibble containing team member names
}
\description{
\code{parse_team} parses queries https://www.procyclingstats.com/teams.php and extracts
team members
}
