% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_team.R
\name{query_team}
\alias{query_team}
\title{Query team membership}
\usage{
query_team(team_names)
}
\arguments{
\item{team_names}{Team name(s) as character vector.}
}
\value{
A tibble containing rider and team names.
}
\description{
A function to query current pro team membership from PCS.
}
\details{
This function is useful to generate a vector of names to use with
\code{query_pcs}. Only works for UCI WorldTeams and ProTeams that exist in the
current year.
}
\examples{

query_team(team_names = c("AG2R Citroen Team","Burgos-BH"))
}
