% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_html_safe.R
\name{read_html_safe}
\alias{read_html_safe}
\title{Get HTML document represented by URL}
\usage{
read_html_safe(url, tries = 3)
}
\arguments{
\item{url}{URL of web page to be downloaded}

\item{tries}{Retry counter (optional, default = 3)}
}
\value{
HTML document, NULL on error
}
\description{
\code{read_html_safe} downloads HTML document from URL address. Function
\code{xml2::read_html} is called. When HTTP error occurs, optional \code{tries}
parameter is decreased and new attempt is performed until \code{tries} counter
reaches zero.
}
