% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rider_profiles_men}
\alias{rider_profiles_men}
\title{Elite men profiles}
\format{
Column description:
\describe{
  \item{\code{rider}}{Rider name}
  \item{\code{dob}}{Date of birth}
  \item{\code{nationality}}{Rider nationality}
  \item{\code{pob}}{Place of birth}
  \item{\code{current_team}}{Current team}
  \item{\code{weight}}{Rider weight in kilograms}
  \item{\code{height}}{Rider height in meters}
  \item{\code{one_day_races}}{Points per "one day races" specialty}
  \item{\code{gc}}{Points per "general classification" specialty}
  \item{\code{tt}}{Points per "time trial" specialty}
  \item{\code{sprint}}{Points per "sprint" specialty}
  \item{\code{climber}}{Points per "climber" specialty}
}
}
\source{
\url{http://www.procyclingstats.com}
}
\usage{
rider_profiles_men
}
\description{
A dataset containing rider profiles of elite men category.
}
\keyword{datasets}
