% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rider_records_men}
\alias{rider_records_men}
\title{Elite men results}
\format{
Column description:
\describe{
  \item{date}{Result date}
  \item{result}{Result (finishing place)}
  \item{gc_result_on_stage}{GC placement if applicable, otherwise NA}
  \item{race}{Name of the race}
  \item{distance}{Covered distance if applicable, otherwise NA}
  \item{pointspcs}{Scored points (PCS ranking)}
  \item{pointsuci}{Scored points (UCI ranking)}
  \item{stage}{Stage description if applicable, 'One day' in case of classic
    races, eventually classification description (e.g. Point classification)}
  \item{rider}{Rider name}
  \item{team}{Current team}
}
}
\usage{
rider_records_men
}
\description{
A dataset containing the results of elite men category.
}
\keyword{datasets}
