% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagan.R
\docType{data}
\name{sagan}
\alias{sagan}
\title{Peter Sagan rider profile and race results through May 2021}
\format{
An object of class \code{list} of length 2.
}
\usage{
sagan
}
\description{
Data scraped from procyclingstats.com using the \code{query_pcs} function.
}
\keyword{datasets}
