% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_pcs.R
\name{query_pcs}
\alias{query_pcs}
\title{Main PCS scraping function}
\usage{
query_pcs(rider_names, seasons = NULL)
}
\arguments{
\item{seasons}{Integer vector of years ("seasons") to collect results from. Will return all years if \code{NULL}.}

\item{rider_urls}{Character vector containing one or more rider names.}
}
\value{
List of two data frames (\code{profiles} and \code{results}).
  See \code{rider_profiles_men} and \code{rider_records_men} documentation
  for details.
}
\description{
\code{query_pcs} scrapes PCS data (rider profiles and results)
for a given vector of rider names.
}
\examples{

# race results for two riders in the 2021 season
query_pcs(c("Peter Sagan","Adam Yates"), seasons = 2021)

# two riders, two seasons
query_pcs(c("Peter Sagan","Adam Yates"), seasons = c(2020,2021))

# two riders, all seasons
query_pcs(c("Peter Sagan","Adam Yates"))
}
