% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rider_profiles_women}
\alias{rider_profiles_women}
\title{Elite women profiles}
\format{
Column description:
\describe{
  \item{rider}{Rider name}
  \item{dob}{Date of birth}
  \item{nationality}{Rider nationality}
  \item{pob}{Place of birth}
  \item{current_team}{Current team}
  \item{weight}{Rider weight in kilograms}
  \item{height}{Rider height in meters}
  \item{one_day_races}{Points per "one day races" specialty}
  \item{gc}{Points per "general classification" specialty}
  \item{tt}{Points per "time trial" specialty}
  \item{sprint}{Points per "sprint" specialty}
  \item{climber}{Points per "climber" specialty}
}
}
\usage{
rider_profiles_women
}
\description{
A dataset containing rider profiles of elite women category.
}
\keyword{datasets}
